+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Changed
+
+- Fix `iframe-has-title` such that if the attribute is ommitted completely that it fires correctly.
+
 ## [0.6.1] - 2021-01-21
 
 ### Changed